pull

+++ peer/reputation.go
@@ -1,5 +1,5 @@
 package peer
 
 type Reputation interface {
-	InitReputationCheck(userID string, reputScore int) error
+	InitReputationCheck(userID string, reputScore int64) error
 }
